Re-arrange controller exports.
Pulls several utils into a new file.
